aider: Change the docker setup to use the "coop.rchain/rnode" with the tag "0.0.1-76-gb0b3545" as the default image.
# Aider chat conversation:

USER: change the docker setup to use the """coop.rchain/rnode""" with the tag """ 0.0.1-76-gb0b3545" as the default image

ASSISTANT: To make this change, we need to modify `shard.yml` and `proposer.yml` to update the default Docker image used by the services.

Here are the *SEARCH/REPLACE* blocks for each file:

shard.yml
```yaml
<<<<<<< SEARCH
x-rnode: &default-rnode
  image: nuttzipper/rnode:schedulers
=======
x-rnode: &default-rnode
  image: coop.rchain/rnode:0.0.1-76-gb0b3545
>>>>>>> REPLACE
```

proposer.yml
```yaml
<<<<<<< SEARCH
services:
  propose:
    build: ./proposerImage
=======
services:
  propose:
    image: coop.rchain/rnode:0.0.1-76-gb0b3545
>>>>>>> REPLACE
```

Please note that in `proposer.yml`, we are replacing the `build` directive with an `image` directive. If the `build` directive is still needed for other purposes, please let me know so we can adjust the changes accordingly.